refactor(@angular-devkit/build-angular): move Angular memory Vite plugin into separate file
To reduce the amount of code within the main Vite development server file, the Vite plugin
that handles the in-memory output file integration for the Angular build system is now within
a separate file.
